% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{as.character.card}
\alias{as.character.card}
\title{Coerce card vector to character}
\usage{
\method{as.character}{card}(x, ...)
}
\arguments{
\item{x}{A card vector}

\item{...}{Additional arguments (ignored)}
}
\value{
A character vector
}
\description{
Returns a character vector like "A♠", "10♦".
}
\examples{
cards <- card(rank = c("A", "10", "Q"), suit = c("♠", "♦", "♥"))
as.character(cards) # [1] "A♠" "10♦" "Q♥"
}
