% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_again.R
\name{ask_play_again}
\alias{ask_play_again}
\title{Ask to play again}
\usage{
ask_play_again()
}
\value{
A logical indicating whether the player wants to play again (TRUE) or exit (FALSE).
}
\description{
Prompts the user to play another round or exit the game.
}
