% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{card}
\alias{card}
\title{Create a card vector}
\usage{
card(suit, rank)
}
\arguments{
\item{suit}{Character vector, allowed values："♠","♥","♦","♣"}

\item{rank}{Character vector, allowed values:："A","2",...,"10","J","Q","K"}
}
\value{
A vctrs record vector with the class "card"
}
\description{
Create a card vector
}
