% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-utils.R
\name{card_is_ace}
\alias{card_is_ace}
\title{judging A (Ace)}
\usage{
card_is_ace(x)
}
\arguments{
\item{x}{A card vector}
}
\value{
Logical vector, TRUE indicates that rank == "A"
}
\description{
judging A (Ace)
}
