% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-utils.R
\name{card_is_face}
\alias{card_is_face}
\title{judging J/Q/K}
\usage{
card_is_face(x)
}
\arguments{
\item{x}{A card vector}
}
\value{
Logical vector, TRUE indicates that the corresponding rank is at c("J","Q","K").
}
\description{
judging J/Q/K
}
