% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{card_rank.blackjack_hand}
\alias{card_rank.blackjack_hand}
\title{card_rank method for blackjack_hand class}
\usage{
\method{card_rank}{blackjack_hand}(x)
}
\arguments{
\item{x}{A blackjack_hand object.}
}
\value{
Character vector of ranks for each card in the hand.
}
\description{
Extracts ranks from the cards contained in a blackjack_hand.
}
