% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_utils.R
\name{card_rank.card}
\alias{card_rank.card}
\title{card_rank method for card class}
\usage{
\method{card_rank}{card}(x)
}
\arguments{
\item{x}{A \code{card} object.}
}
\value{
Character vector of ranks.
}
\description{
Extracts the rank field from a \code{card} object.
}
\keyword{internal}
