% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_utils.R
\name{card_rank.default}
\alias{card_rank.default}
\title{Default method for card_rank}
\usage{
\method{card_rank}{default}(x)
}
\arguments{
\item{x}{Object of any class.}
}
\description{
Returns \code{NULL} for unsupported classes.
}
