% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_utils.R
\name{card_suit}
\alias{card_suit}
\title{Extract the suit from a card vector}
\usage{
card_suit(x)
}
\arguments{
\item{x}{A card vector}
}
\value{
Character vectors, such as "♠","♥","♦","♣"
}
\description{
Extract the suit from a card vector
}
\examples{
cards <- as_card(c("A♠", "10♦", "Q♥"))
card_suit(cards)
}
