% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-utils.R
\name{card_value}
\alias{card_value}
\title{Obtain the initial points of a single card in Blackjack
(ensure J/Q/K == 10，A == 11)}
\usage{
card_value(x)
}
\arguments{
\item{x}{A card vector}
}
\value{
Integer vector
}
\description{
Obtain the initial points of a single card in Blackjack
(ensure J/Q/K == 10，A == 11)
}
