% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealer_action.R
\name{dealer_action}
\alias{dealer_action}
\title{Dealer action logic}
\usage{
dealer_action(hand, deck)
}
\arguments{
\item{hand}{A `blackjack_hand` object}

\item{deck}{A character vector of remaining cards}
}
\value{
A list with updated `hand` and `deck`
}
\description{
Simulate dealer's automatic behavior: keep hitting until score >= 17.
}
