% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_loop.R
\name{exit}
\alias{exit}
\title{Exit Blackjack session}
\usage{
exit()
}
\value{
Invisibly returns NULL.
}
\description{
Intended for ending an interactive Blackjack session without using play().
}
\examples{
exit()  # Prints "Exiting Blackjack session. Goodbye!"
}
