% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{format.blackjack_hand}
\alias{format.blackjack_hand}
\title{Format a Blackjack Hand}
\usage{
\method{format}{blackjack_hand}(x, ...)
}
\arguments{
\item{x}{A `blackjack_hand` object}

\item{...}{Additional arguments (ignored)}
}
\value{
A character string representing the hand
}
\description{
Custom print method for `blackjack_hand` objects.
}
