% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{format.card}
\alias{format.card}
\title{Format a card vector}
\usage{
\method{format}{card}(x, ...)
}
\arguments{
\item{x}{A card vector}

\item{...}{Additional arguments passed to or from other methods}
}
\value{
A character vector
}
\description{
Returns a character representation like "A♠", "10♦".
}
\keyword{internal}
