% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_utils.R
\name{new_blackjack_hand}
\alias{new_blackjack_hand}
\title{Create a blackjack hand (S3 vctrs)}
\usage{
new_blackjack_hand(cards)
}
\arguments{
\item{cards}{Character vector of card values, e.g. c("A", "10", "5")}
}
\value{
A `blackjack_hand` S3 object
}
\description{
Create a hand object representing a set of cards
}
