% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_data.R
\docType{data}
\name{players_data}
\alias{players_data}
\title{Sample player data for vignette examples}
\format{
A list with components:
\describe{
  \item{players}{players in the dataset}
  \item{bankroll_history}{A data frame with the bankroll history of each player}
  }
}
\source{
Internal example data for wombat21
}
\usage{
players_data
}
\description{
A list containing example players and bankroll history.
}
\keyword{internal}
