% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{print.card}
\alias{print.card}
\title{Print a card vector}
\usage{
\method{print}{card}(x, ...)
}
\arguments{
\item{x}{A card vector}

\item{...}{Additional arguments passed to or from other methods}
}
\description{
Nicely prints a card vector as "A♠ 10♦ Q♥".
}
\examples{
cards <- card(rank = c("A", "10", "Q"), suit = c("♠", "♦", "♥"))
print(cards)
}
