% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{reset_game_data}
\alias{reset_game_data}
\title{Reset game memory: players_db and/or bankroll_history}
\usage{
reset_game_data(
  players_db,
  bankroll_history,
  reset_players = TRUE,
  reset_history = TRUE
)
}
\arguments{
\item{players_db}{Player Database in Local Directory}

\item{bankroll_history}{List tracking coin balances over rounds.}

\item{reset_players}{Logical. Whether to clear players_db}

\item{reset_history}{Logical. Whether to clear bankroll_history}
}
\value{
A list with (possibly reset) players_db and bankroll_history
}
\description{
Reset game memory: players_db and/or bankroll_history
}
