% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\docType{data}
\name{simulation_data}
\alias{simulation_data}
\title{Sample simulation data for vignette examples}
\format{
A list with components:
\describe{
  \item{num_players}{Number of players in the simulation.}
  \item{num_rounds}{Number of rounds simulated}
  \item{players}{A data frame containing player statistics, including: win rate, lose rate, tie rate, insurance win rate, split rate, split win, split lose, final #' coins, bankroll history, initial coins, bet amount and threshold}
  \item{overall}{A data frame with overall statistics for the simulation}
  \item{bankroll_history}{A data frame with the bankroll history of each player}
  }
}
\source{
Internal example data for wombat21
}
\usage{
simulation_data
}
\description{
A list containing example players and bankroll history.
}
\keyword{internal}
