% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{vec_cast.card.character}
\alias{vec_cast.card.character}
\title{Cast character vector to card}
\usage{
vec_cast.card.character(x, to, ...)
}
\arguments{
\item{x}{A character vector}

\item{to}{A card prototype}
}
\value{
A card vector
}
\description{
Converts character like "A♠", "10♦" into a card object.
}
