% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{vec_cast.card}
\alias{vec_cast.card}
\title{Cast anything to card (identity or error)}
\usage{
\method{vec_cast}{card}(x, to, ...)
}
\arguments{
\item{x}{Input object.}

\item{to}{Card prototype.}

\item{...}{Additional arguments (ignored)}
}
\value{
Card vector or error
}
\description{
Used internally by vctrs for S3 casting.
}
\keyword{internal}
