% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{vec_cast.character.card}
\alias{vec_cast.character.card}
\title{Cast card to character vector}
\usage{
vec_cast.character.card(x, to, ...)
}
\arguments{
\item{x}{A card object}

\item{to}{Unused}
}
\value{
Character vector
}
\description{
Converts a card object back to character like "Q♥".
}
