% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{vec_ptype2.card.card}
\alias{vec_ptype2.card.card}
\title{Type promotion: card + card → card prototype}
\usage{
\method{vec_ptype2}{card.card}(x, y, ...)
}
\description{
Type promotion: card + card → card prototype
}
