% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{vec_ptype2.card.character}
\alias{vec_ptype2.card.character}
\title{Type promotion method for card objects}
\usage{
\method{vec_ptype2}{card.character}(x, y, ...)
}
\arguments{
\item{x}{A card or character vector}

\item{y}{Another card or character vector}

\item{...}{Additional arguments passed to or from other methods}
}
\value{
A card prototype
}
\description{
Implements method dispatch for combining card and character/card types.
}
