% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{vec_ptype2.character.card}
\alias{vec_ptype2.character.card}
\title{Type promotion: character + card → card}
\usage{
vec_ptype2.character.card(x, y, ...)
}
\description{
Type promotion: character + card → card
}
