% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cal.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate Blackjack Score}
\usage{
calculate_score(hand)
}
\arguments{
\item{hand}{A `blackjack_hand` object}
}
\value{
Numeric score of the hand (integer)
}
\description{
Calculates the total point value of a blackjack hand using C++
}
