% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{card}
\alias{card}
\title{Create a card vector}
\usage{
card(rank = character(), suit = character())
}
\arguments{
\item{rank}{Character vector, allowed values:："A","2",...,"10","J","Q","K"}

\item{suit}{Character vector, allowed values："♠","♥","♦","♣"}
}
\value{
A vctrs record vector with the class "card"
}
\description{
Create a card vector
}
\examples{
cards <- card(rank = c("A", "10", "Q"), suit = c("♠", "♦", "♥"))
print(cards)
}
