% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{card}
\alias{card}
\title{Create a card vector}
\usage{
card(rank = character(), suit = character())
}
\arguments{
\item{rank}{Character vector, allowed values: "A","2",...,"10","J","Q","K"}

\item{suit}{Character vector, allowed values: "♠","♥","♦","♣"}
}
\value{
A vctrs record vector with class "card"
}
\description{
Constructs a custom card object using vctrs::new_rcrd().
}
