% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_winner.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine winner(s) of a Blackjack game}
\usage{
determine_winner(player_hand, dealer_hand)
}
\arguments{
\item{player_hand}{A `blackjack_hand` or list of `blackjack_hand`}

\item{dealer_hand}{A `blackjack_hand` object}
}
\value{
Outcome string or named character vector if multiple hands
}
\description{
Handles single hand or list of hands; supports Blackjack, Charlie, Split
}
