% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ending.R
\name{end_round}
\alias{end_round}
\title{End round processing}
\usage{
end_round(player_hands, dealer_hand, players, players_db, bankroll_history)
}
\arguments{
\item{player_hands}{Named list of player hands.}

\item{dealer_hand}{Dealer's hand.}

\item{players}{Named list of players with coin balances.}

\item{players_db}{Player Database in Local Directory}

\item{bankroll_history}{List tracking coin balances over rounds.}
}
\value{
A list with updated `players` and `bankroll_history`.
}
\description{
Handles end of round operations: displaying results, saving players,
updating bankroll history, and plotting results.
}
