% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{handle_splitting}
\alias{handle_splitting}
\title{Handle splitting}
\usage{
handle_splitting(player_hands, deck, players)
}
\arguments{
\item{player_hands}{A list of lists of blackjack_hand objects.}

\item{deck}{A character vector representing the remaining deck.}

\item{players}{A named list of player objects (including names and is_computer flag).}
}
\value{
An updated list of lists of blackjack_hand objects after potential splits.
}
\description{
Allows players with pairs to split their hand.
}
