% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_player.R
\name{input_players}
\alias{input_players}
\title{Input and validate human and computer players}
\usage{
input_players(players_db)
}
\arguments{
\item{players_db}{Player Database in Local Directory}
}
\value{
A list with two elements:
  \describe{
    \item{session_players}{A named list of players (human and computer) with their info for the current session.}
    \item{players_db}{The updated player database (named list).}
  }
}
\description{
Allows input of human player names, creates computer players,
             loads or creates player data with coins.
}
\examples{
\dontrun{
# Typical usage (prompts user for input):
res <- input_players(players_db = list())
str(res$session_players)
}
}
