% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_player.R
\name{input_players}
\alias{input_players}
\title{Input and validate human and computer players}
\usage{
input_players(players_db)
}
\arguments{
\item{players_db}{Player Database in Local Directory}
}
\value{
A named list of players (human + computer), where each element is a list with player info (name, money, is_computer).
}
\description{
Allows input of human player names, creates computer players, loads or creates player data with coins.
}
