% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_type.R
\name{is_five_card_charlie}
\alias{is_five_card_charlie}
\title{Check 5-card Charlie condition}
\usage{
is_five_card_charlie(hand)
}
\arguments{
\item{hand}{A `blackjack_hand` object}
}
\value{
TRUE if qualifies as 5-card Charlie, else FALSE
}
\description{
Returns TRUE if the hand has 5 or more cards and score ≤ 21
}
\examples{
# Example: 5-card Charlie
hand <- new_blackjack_hand(c("2♠", "3♥", "4♦", "5♣", "7♠"))
is_five_card_charlie(hand)
}
