% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_player.R
\name{load_players}
\alias{load_players}
\title{Load players from disk}
\usage{
load_players()
}
\value{
A list containing players and bankroll_history, or default empty lists if none found
}
\description{
Load players from disk
}
