% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_players}
\alias{save_players}
\title{Save players to disk}
\usage{
save_players(players, bankroll_history = NULL)
}
\arguments{
\item{players}{named list of players}

\item{bankroll_history}{A list. Each element is a named numeric vector of player coins per round.}
}
\description{
Save players to disk
}
