% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-class.R
\name{vec_ptype2.card.character}
\alias{vec_ptype2.card.character}
\title{Type promotion: card + character → card}
\usage{
vec_ptype2.card.character(x, y, ...)
}
\description{
Type promotion: card + character → card
}
